Various general improvements and server status additons.
Determined the cause and fixed the issue with the database connection. Just didnt have the database in the right location... Fixed that so that now the database is copied to the correct location on build.

Added some basic server functionality to both the database and commands. The database can now hold a server entry that contains the executable location and process name so that commands can now be used to determine both server status and start the server.

Also added a help command that is auto-generated by the commands instance available to the general commands class.